requirements.txt
google-api-python-client>=2.172.0
streamlit==1.45.0
langchain-huggingface==0.1.1
langchain-community==0.3.23
groq==0.18.0
sentence-transformers==4.1.0
faiss-cpu==1.11.0
pandas==2.2.2
plotly==5.21.0
requests==2.32.2
python-dotenv==1.0.1
pymongo==4.7.2
pytest==8.3.4
